Fixed divider and BreadcrumbItem stories
