+++ packages/react-components/react-breadcrumb-preview/stories/Breadcrumb/BreadcrumbFocusMode.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { Breadcrumb, BreadcrumbItem, BreadcrumbButton, BreadcrumbDivider } from '@fluentui/react-breadcrumb-preview';
-import { bundleIcon, CalendarMonth20Filled, CalendarMonth20Regular, GridDots20Regular } from '@fluentui/react-icons';
+import { bundleIcon, CalendarMonth20Filled, CalendarMonth20Regular } from '@fluentui/react-icons';
 import { ButtonProps } from '@fluentui/react-components';
 
 const CalendarMonth = bundleIcon(CalendarMonth20Filled, CalendarMonth20Regular);
@@ -10,9 +10,7 @@ type Item = {
   item?: string;
   buttonProps?: {
     'aria-label'?: string;
-    onClick?: () => void;
     icon?: ButtonProps['icon'];
-    disabled?: boolean;
   };
 };
 
@@ -32,24 +30,10 @@ const buttonItems: Item[] = [
   {
     key: 2,
     item: 'Item 2',
-    buttonProps: {
-      icon: <GridDots20Regular />,
-    },
   },
   {
     key: 3,
     item: 'Item 3',
-  },
-  {
-    key: 4,
-    item: 'Item 4',
-    buttonProps: {
-      icon: <CalendarMonth />,
-    },
-  },
-  {
-    key: 5,
-    item: 'Item 5',
   },
 ];
 
@@ -57,11 +41,7 @@ function renderButton(el: Item, isLastItem: boolean = false) {
   return (
     <React.Fragment key={`${el.key}-button`}>
       <BreadcrumbItem>
-        <BreadcrumbButton
-          {...el.buttonProps}
-          current={isLastItem}
-          onClick={isLastItem ? undefined : el.buttonProps?.onClick}
-        >
+        <BreadcrumbButton {...el.buttonProps} current={isLastItem}>
           {el.item}
         </BreadcrumbButton>
       </BreadcrumbItem>
@@ -69,17 +49,8 @@ function renderButton(el: Item, isLastItem: boolean = false) {
     </React.Fragment>
   );
 }
-export const BreadcrumbFocusMode = () => (
-  <>
-    <h2>`tab` - default</h2>
-    Navigation using `tab` key.
-    <Breadcrumb aria-label="Breadcrumb with `tab` focusMode">
-      {buttonItems.map(el => renderButton(el, el.key === buttonItems.length - 1))}
-    </Breadcrumb>
-    <h2>`arrow`</h2>
-    Navigation using arrow keys
-    <Breadcrumb aria-label="Breadcrumb with `arrow` focusMode" focusMode="arrow">
-      {buttonItems.map(el => renderButton(el, el.key === buttonItems.length - 1))}
-    </Breadcrumb>
-  </>
+export const FocusModeArrow = () => (
+  <Breadcrumb aria-label="Breadcrumb with `arrow` focusMode" focusMode="arrow">
+    {buttonItems.map(el => renderButton(el, el.key === buttonItems.length - 1))}
+  </Breadcrumb>
 );
