--- packages/react-components/react-breadcrumb-preview/stories/BreadcrumbButton/BreadcrumbButtonBestPractices.md
@@ -1,12 +1,0 @@
-<details>
-<summary>
- Best Practices
-</summary>
-
-### Do
-
-- Put `current` attribute to the last item.
-- Use `aria-current="page"` to indicate which page is currently displayed.
-- Use descriptive `aria-label` for BreadcrumbButton which contains only icon.
-
-</details>
