--- packages/react-components/react-breadcrumb-preview/stories/BreadcrumbButton/BreadcrumbButtonDescription.md
@@ -1,1 +1,0 @@
-BreadcrumbButton uses @fluentui/react-button under the hood.
