+++ packages/react-components/react-breadcrumb-preview/stories/BreadcrumbButton/BreadcrumbButtonWithHrefAttribute.stories.tsx
@@ -1,11 +1,5 @@
 import * as React from 'react';
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbButton,
-  BreadcrumbDivider,
-  BreadcrumbProps,
-} from '@fluentui/react-breadcrumb-preview';
+import { Breadcrumb, BreadcrumbItem, BreadcrumbButton, BreadcrumbDivider } from '@fluentui/react-breadcrumb-preview';
 import { bundleIcon, CalendarMonth20Filled, CalendarMonth20Regular, GridDots20Regular } from '@fluentui/react-icons';
 import { ButtonProps } from '@fluentui/react-components';
