--- packages/react-components/react-breadcrumb-preview/stories/BreadcrumbDivider/BreadcrumbDividerBestPractices.md
@@ -1,14 +1,0 @@
-<details>
-<summary>
- Best Practices
-</summary>
-
-### Do
-
-- Use `slash` dividers only for small and non-interactive breadcrums. Use it to describe file paths.
-
-### Don't
-
-- Avoid using custom dividers.
-
-</details>
