--- packages/react-components/react-breadcrumb-preview/stories/BreadcrumbDivider/BreadcrumbDividerDescription.md
@@ -1,1 +1,0 @@
-A divider component which separates breadcrumb items.
