--- packages/react-components/react-breadcrumb-preview/stories/BreadcrumbItem/BreadcrumbItemDescription.md
@@ -1,1 +1,0 @@
-BreadcrumbItem can be used as non-interactive element or to be a container for BreadcrumbButton and BreadcrumbLink components.
